+++ src/main/java/com/controllers/LoadUserController.java
@@ -1,5 +1,6 @@
 package com.controllers;
 
+import com.helpers.AlertHelper;
 import com.helpers.UserHelper;
 import com.models.User;
 import com.typingtest.Main;
@@ -32,7 +33,8 @@ public class LoadUserController {
                     Header.usernameLabel.setText(userSelected.getName());
                     Header.highWPM.setText("Highest WPM: " + userSelected.getBestWPM());
                     Header.avgWPMLabel.setText("Average WPM: " + userSelected.getAverageWPM());
-
+                    Header.userTooltip.setText("Highest WPM: " + userSelected.getBestWPM() +
+                                                "\nAverage WPM: " + userSelected.getAverageWPM());
                     LoadUser.getStage().close();
                 }
             }
@@ -44,12 +46,16 @@ public class LoadUserController {
             @Override
             public void handle(ActionEvent actionEvent) {
                 int itemIndex = LoadUser.users.getSelectionModel().getSelectedIndex();
+
+                User userSelected = UserHelper.getUser(itemIndex);
+                boolean userConfirmation = AlertHelper.createAlertConformation("Delete User Confirmation",
+                        "Are you sure you want to delete '" + userSelected.getName()+ "'?");
+                if(!userConfirmation) return;
+
                 if(Main.primaryScene.getRoot() == TestResults.getBorder()) {
                     TitleScreen.getBorder().setTop(Header.getHeader());
                     Main.primaryScene.setRoot(TitleScreen.getBorder());
                 }
-
-                User userSelected = UserHelper.getUser(itemIndex);
                 UserHelper.deleteUser(userSelected);
                 UserHelper.addUsersToList();
             }
